remove now deprecated methods, improve SidecarWatcher usability, cleanup
